refactor(mbx): split CLI commands into modules
